move mhv-related client methods into clients
