+++ lib/rx/client.rb
@@ -8,15 +8,10 @@ require 'common/client/middleware/response/snakecase'
 require 'rx/middleware/response/rx_parser'
 require 'rx/configuration'
 require 'rx/client_session'
-require 'rx/api/prescriptions'
-require 'rx/api/sessions'
 
 module Rx
   # Core class responsible for api interface operations
   class Client
-    include Rx::API::Prescriptions
-    include Rx::API::Sessions
-
     REQUEST_TYPES = %i(get post).freeze
     USER_AGENT = 'Vets.gov Agent'
     BASE_REQUEST_HEADERS = {
@@ -38,6 +33,46 @@ module Rx
         @session.save
       end
       self
+    end
+
+    def get_active_rxs
+      json = perform(:get, 'prescription/getactiverx', nil, token_headers).body
+      Common::Collection.new(::Prescription, json)
+    end
+
+    def get_history_rxs
+      json = perform(:get, 'prescription/gethistoryrx', nil, token_headers).body
+      Common::Collection.new(::Prescription, json)
+    end
+
+    def get_rx(id)
+      collection = get_history_rxs
+      collection.find_first_by('prescription_id' => { 'eq' => id })
+    end
+
+    def get_tracking_rx(id)
+      json = perform(:get, "prescription/rxtracking/#{id}", nil, token_headers).body
+      data = json[:data].first.merge(prescription_id: id)
+      Tracking.new(json.merge(data: data))
+    end
+
+    def get_tracking_history_rx(id)
+      json = perform(:get, "prescription/rxtracking/#{id}", nil, token_headers).body
+      tracking_history = json[:data].map { |t| Hash[t].merge(prescription_id: id) }
+      Common::Collection.new(::Tracking, json.merge(data: tracking_history))
+    end
+
+    def post_refill_rx(id)
+      perform(:post, "prescription/rxrefill/#{id}", nil, token_headers)
+    end
+
+    def get_session
+      env = perform(:get, 'session', nil, auth_headers)
+      req_headers = env.request_headers
+      res_headers = env.response_headers
+      Rx::ClientSession.new(user_id: req_headers['mhvCorrelationId'],
+                            expires_at: res_headers['expires'],
+                            token: res_headers['token'])
     end
 
     private
